migrate anon funcs to utils
